backend/admin_tools/requirements.txt
google
google-cloud
google-cloud-api-keys
google-cloud-firestore
google-cloud-texttospeech
google-cloud-storage
google-api-python-client
google-genai
google-adk
python-pptx
pandas
openpyxl
PyYAML